backend/requirements.txt
Django==5.1.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-storages==1.14.4
boto3==1.35.76
reportlab==4.2.5
python-dotenv==1.0.1
django-cors-headers==4.6.0
dj-database-url==2.3.0
